New option to specify tables to skip via a file.
The --regex option works well but cannot scale to hundreds of thousands
of distinct table names when a large number of tables must not be backed
up.

This new --omit-from-file option allows specifying a file containing a
list of table names (in database.table format, one per line) to skip
during backup.