Remove redundant OpenSSL check
mydumper itself doesn't use OpenSSL, libmysqlclient does. Checking to
see if OpenSSL is installed on the machine compiling mydumper does not
check if libmysqlclient was built with OpenSSL. libmysqlclient could for
example have been build with OpenSSL statically compiled. Also MySQL 8.0
doesn't have this as an option and MariaDB Connector/C can compile with
other SSL libraries.

If the intention is to check of libmysqlclient was built with OpenSSL
then a symbol check should be made instead.