validate thread count cli options
there does not appear to be validation of the --threads or -t count
passed into mydumper.  Invalid options like -1 cause the program
to crash and valid but low thread counts like 0 or 1 cause the
program to hang indefinately and in the case of 1 with a consistent
backup cause the database to lock until the program is killed.  This
change is a very simple approach to validating that the thread count
is at least two or a warning is printed and the count set to 2.  Invalid
types passed in like signed numbers will still cause mydumper to crash
but not hang and potentially lock the database.