mydumper fails to restore view for table with 1026 columns
The logic in write_view_definition_into_file() of making duplicate
CREATE TABLE statements for views commented in the code as "we create
tables as workaround for view dependencies" pursues the goal of fixing
Bug#10927.

But this was done with undefined storage engine. Default storage
engine prohibits such amount of columns. We add ENGINE=MEMORY which
allows such amount of columns.

Original fix for Bug#10927 adds TEMPORARY keyword but we cannot add it
as it will fail for multiple connections.

Bug description:

https://jira.mariadb.org/browse/CONTRIB-14