macOS: minor compiler compatibility (#2139)
This patch fixes the issue with an unused and unusable arg getting
passed to a function that does not strictly define its args list.  This
was causing CLANG to fail compilation.

Tested on "manually" on macOS against mariadb 10.6.16 and mysql 8.4. The
full test suite was run on debian against mariadb and mysql.